refactor: move theme hook to useTheme file
