+++ src/App.jsx
@@ -22,10 +22,10 @@ import Register from "./screens/Register/Register.jsx";
 import SelectionScreen from "./screens/SelectionScreen/SelectionScreen.jsx";
 import Settings from "./screens/Settings/Settings.jsx";
 
+import useTheme from "./hooks/useTheme.js";
 import { setLanguages } from "./redux/Actions/language.js";
 import { signIn, signOut } from "./redux/Actions/login.js";
 import { getLanguages, getProfile } from "./utils/api.js";
-import { useTheme } from "./utils/theme.js";
 
 import Image from "./images/logo/color-round.svg";
